feat(mokumo-shop): move MokumoApp Graft impl from services/api [#512]
MokumoApp Graft impl now lives in crates/mokumo-shop/src/graft.rs:
- type AppState = SharedMokumoState (Arc<MokumoState>)
- type DomainState = MokumoShopState
- build_domain_state() constructs MokumoShopState (ws, local_ip,
  restore_in_progress, restore_limiter, ws_ping_ms)
- compose_state() assembles MokumoState from three slices
- platform_state() / control_plane_state() return field references

services/api/src/graft.rs becomes a re-export shim. MokumoAppState
remains in services/api as the transitional construction type until
build_app_inner is replaced by Engine::boot() in Session 2.3.
